[feature/#928] fix to AuthCertificationActivity
